Separate logger setup for production and non-production environments
